Extracting extract_tag from deploy.
This completely removes verify_deployments' dependency from deploy.py.
